generate example data sets local_code and exported
Prior to this change the `local_code` object was not exported and thus not
available for the examples in the documentation.

This change moved the `local_code` and `local_code_regular_assignment`
construction to a script in the data-raw directory and exports the two
objects as data sets in the package.

re EvaMaeRey/flipbookr#5